+++ libp2p/security/insecure/transport.py
@@ -1,13 +1,12 @@
+from typing import cast
+
 from libp2p.security.secure_transport_interface import ISecureTransport
 from libp2p.security.secure_conn_interface import ISecureConn
 
-from typing import TYPE_CHECKING, Dict, Any, cast
 
-if TYPE_CHECKING:
-    from libp2p.network.connection.raw_connection_interface import IRawConnection
-    from libp2p.peer.id import ID
-    from .secure_conn_interface import ISecureConn
-    from .typing import TSecurityDetails
+from libp2p.network.connection.raw_connection_interface import IRawConnection
+from libp2p.peer.id import ID
+from libp2p.security.typing import TSecurityDetails
 
 
 class InsecureTransport(ISecureTransport):
@@ -21,7 +20,7 @@ class InsecureTransport(ISecureTransport):
     def __init__(self, transport_id: str) -> None:
         self.transport_id = transport_id
 
-    async def secure_inbound(self, conn: "IRawConnection") -> ISecureConn:
+    async def secure_inbound(self, conn: IRawConnection) -> ISecureConn:
         """
         Secure the connection, either locally or by communicating with opposing node via conn,
         for an inbound connection (i.e. we are not the initiator)
@@ -30,9 +29,7 @@ class InsecureTransport(ISecureTransport):
         insecure_conn = InsecureConn(conn, self.transport_id)
         return insecure_conn
 
-    async def secure_outbound(
-        self, conn: "IRawConnection", peer_id: "ID"
-    ) -> ISecureConn:
+    async def secure_outbound(self, conn: IRawConnection, peer_id: ID) -> ISecureConn:
         """
         Secure the connection, either locally or by communicating with opposing node via conn,
         for an inbound connection (i.e. we are the initiator)
@@ -43,21 +40,21 @@ class InsecureTransport(ISecureTransport):
 
 
 class InsecureConn(ISecureConn):
-    conn: "IRawConnection"
-    details: "TSecurityDetails"
+    conn: IRawConnection
+    details: TSecurityDetails
 
-    def __init__(self, conn: "IRawConnection", conn_id: str) -> None:
+    def __init__(self, conn: IRawConnection, conn_id: str) -> None:
         self.conn = conn
-        self.details = cast("TSecurityDetails", {})
-        self.details["id"] = conn_id
+        self.details = cast(TSecurityDetails, {})
+        self.details[id] = conn_id
 
-    def get_conn(self) -> "IRawConnection":
+    def get_conn(self) -> IRawConnection:
         """
         :return: connection object that has been made secure
         """
         return self.conn
 
-    def get_security_details(self) -> "TSecurityDetails":
+    def get_security_details(self) -> TSecurityDetails:
         """
         :return: map containing details about the connections security
         """
